+++ frontend/src/index.js
@@ -2,9 +2,9 @@ import React from 'react';
 import ReactDOM from 'react-dom/client';
 import './index.css';
 import App from './App';
-import reportWebVitals from './reportWebVitals';
+// import reportWebVitals from './reportWebVitals';
 import { Provider } from 'react-redux';
-import {store} from './app/Store'
+import {store} from './redux/store/Store'
 import { ReactFlowProvider } from 'reactflow';
 
 const root = ReactDOM.createRoot(document.getElementById('root'));
@@ -23,4 +23,4 @@ root.render(
 // If you want to start measuring performance in your app, pass a function
 // to log results (for example: reportWebVitals(console.log))
 // or send to an analytics endpoint. Learn more: https://bit.ly/CRA-vitals
-reportWebVitals();
+// reportWebVitals();
