+++ frontend/src/pages/About.js
@@ -7,7 +7,7 @@ const About = () => {
   <div className="container m-auto px-6 text-gray-600 md:px-12 xl:px-6">
       <div className="space-y-6 md:space-y-0 md:flex md:gap-6 lg:items-center lg:gap-12">
         <div className="md:5/12 lg:w-5/12">
-          <img src="https://images.unsplash.com/photo-1617791160536-598cf32026fb?q=80&w=1964&auto=format&fit=crop&ixlib=rb-4.0.3&ixid=M3wxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8fA%3D%3D" alt="image" loading="lazy" width="" height="" className='rounded-md drop-shadow-md hover:scale-105 ease-linear blur-sm hover:blur-none '/>
+          <img src="https://images.unsplash.com/photo-1617791160536-598cf32026fb?q=80&w=1964&auto=format&fit=crop&ixlib=rb-4.0.3&ixid=M3wxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8fA%3D%3D" alt="image" loading="lazy" width="" height="" className='cursor-pointer rounded-md drop-shadow-md hover:scale-105 ease-linear blur-sm hover:blur-none '/>
         </div>
         <div className="md:7/12 lg:w-6/12">
           <h2 className="text-2xl text-gray-900  dark:text-gray-50 font-bold md:text-4xl">Unleash the power of your mind with our interactive Mind Mapping tool</h2>
