+++ frontend/src/redux/actions/hamburgerslice.js
@@ -1,0 +1,15 @@
+import { createSlice } from "@reduxjs/toolkit";
+
+const initialState =false
+const slice =createSlice({
+    name:"hamburgertoggler",
+    initialState,
+    reducers:{
+        ison:(state,action)=>{
+            return !state
+        }
+    }
+}
+)
+export const  {ison} =slice.actions
+export default slice.reducer
